+++ packages/ui/CHANGELOG.md
@@ -12,6 +12,8 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Fixed
 
+- `ui/PlaceholderDecorator` to update bounds of `Scroller` when the `visible` state changed.
+
 ## [1.4.1] - 2017-07-05
 
 No significant changes.